Revert "Merge pull request #1070 from mzazrivec/schedule_migration_modal_changes_for_warm_migration"
This reverts commit 8c5ac6dec79596391fdfb60e650665b1c39d7e06, reversing
changes made to 83242bd8aba5819c89619f09dc91a47a987f0eea.
