+++ app/javascript/react/screens/App/Overview/OverviewActions.js
@@ -197,40 +197,6 @@ export const toggleScheduleMigrationModal = plan => ({
   payload: plan
 });
 
-export const scheduleCutover = payload => dispatch =>
-  dispatch({
-    type: V2V_SCHEDULE_MIGRATION,
-    payload: new Promise((resolve, reject) => {
-      const {
-        cutoverTime,
-        plan: { id: planId }
-      } = payload;
-      const url = `/api/service_templates/${planId}`;
-      const body = {
-        action: 'edit',
-        resource: {
-          config_info: {
-            warm_migration_cutover_datetime: cutoverTime
-          }
-        }
-      };
-      return API.post(url, body)
-        .then(response => {
-          resolve(response);
-          dispatch({
-            type: V2V_NOTIFICATION_ADD,
-            message: sprintf(
-              __('Migration cutover successfully scheduled for %s'),
-              formatDateTime(response.data.options.config_info.warm_migration_cutover_datetime)
-            ),
-            notificationType: 'success',
-            actionEnabled: false
-          });
-        })
-        .catch(e => reject(e));
-    })
-  });
-
 export const scheduleMigration = payload => dispatch =>
   dispatch({
     type: V2V_SCHEDULE_MIGRATION,
