+++ app/javascript/react/screens/App/Overview/components/ScheduleMigrationModal/ScheduleMigrationModal.js
@@ -1,20 +1,11 @@
-import moment from 'moment';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Button, Modal } from 'patternfly-react';
 import ScheduleMigrationModalBody from './ScheduleMigrationModalBody';
 import getPlanScheduleInfo from '../Migrations/helpers/getPlanScheduleInfo';
-import isPlanWarmMigration from '../Migrations/helpers/isPlanWarmMigration';
 
 class ScheduleMigrationModal extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      dateTimeInput: null,
-      startMigrationNow: null
-    };
-  }
+  state = { dateTimeInput: '' };
 
   render() {
     const {
@@ -23,67 +14,28 @@ class ScheduleMigrationModal extends React.Component {
       scheduleMigrationPlan,
       scheduleMigration,
       fetchTransformationPlansAction,
-      fetchTransformationPlansUrl,
-      scheduleMigrationNow,
-      scheduleCutover
+      fetchTransformationPlansUrl
     } = this.props;
 
-    const handleDatepickerChange = event => {
-      this.setState({ dateTimeInput: event });
-    };
+    const { migrationScheduled } = getPlanScheduleInfo(scheduleMigrationPlan);
 
-    const setScheduleMode = event => {
-      this.setState({ startMigrationNow: event });
+    const handleChange = event => {
+      this.setState({ dateTimeInput: event });
     };
 
     const modalClose = () => {
       toggleScheduleMigrationModal();
+      handleChange();
     };
-
-    const modalMode = plan => {
-      let mode = 'migration';
-      if (isPlanWarmMigration(plan) && plan.status === 'Ok') {
-        mode = 'cutover';
-      } else if (plan != null && (plan.status === 'Error' || plan.status === 'Denied')) {
-        mode = 'retry';
-      }
-      return mode;
-    };
-
-    const MODAL_STRINGS = {
-      migration: [
-        __('Schedule Migration'),
-        __('Start migration immediately'),
-        __('Select date and time for the start of the migration')
-      ],
-      cutover: [__('Schedule Cutover'), __('Start cutover immediately'), __('Select date and time to start cutover')],
-      retry: [__('Schedule Retry'), __('Retry migration immediately'), __('Select date and time to retry migration')]
-    };
-
-    const modalTitle = plan => MODAL_STRINGS[modalMode(plan)][0];
-
-    const modalBodyStrings = plan => [MODAL_STRINGS[modalMode(plan)][1], MODAL_STRINGS[modalMode(plan)][2]];
-
-    const { migrationScheduled, migrationCutover } = getPlanScheduleInfo(scheduleMigrationPlan);
-    const defaultDate = isPlanWarmMigration(scheduleMigrationPlan)
-      ? new Date(migrationCutover)
-      : new Date(migrationScheduled);
 
     return (
       <Modal show={scheduleMigrationModal} onHide={modalClose} backdrop="static">
         <Modal.Header>
           <Modal.CloseButton onClick={modalClose} />
-          <Modal.Title>{modalTitle(scheduleMigrationPlan)}</Modal.Title>
+          <Modal.Title>{__('Schedule Migration Plan')}</Modal.Title>
         </Modal.Header>
         <Modal.Body>
-          <ScheduleMigrationModalBody
-            handleDatepickerChange={handleDatepickerChange}
-            setScheduleMode={setScheduleMode}
-            defaultDate={defaultDate}
-            startNowLabel={modalBodyStrings(scheduleMigrationPlan)[0]}
-            startLaterLabel={modalBodyStrings(scheduleMigrationPlan)[1]}
-            showDatepicker={!this.state.startMigrationNow}
-          />
+          <ScheduleMigrationModalBody handleChange={handleChange} defaultDate={migrationScheduled || ''} />
         </Modal.Body>
         <Modal.Footer>
           <Button bsStyle="default" className="btn-cancel" onClick={modalClose}>
@@ -91,38 +43,18 @@ class ScheduleMigrationModal extends React.Component {
           </Button>
           <Button
             bsStyle="primary"
+            disabled={!this.state.dateTimeInput}
             onClick={() => {
-              if (modalMode(scheduleMigrationPlan) === 'cutover') {
-                let cutoverDatetime = this.state.dateTimeInput;
-                if (this.state.startMigrationNow) {
-                  // now == schedule warm migration cutover 2 minutes from now
-                  cutoverDatetime = moment(new Date())
-                    .add(2, 'm')
-                    .toDate();
-                }
-                scheduleCutover({
-                  plan: scheduleMigrationPlan,
-                  cutoverTime: cutoverDatetime
-                }).then(() => {
-                  fetchTransformationPlansAction({
-                    url: fetchTransformationPlansUrl,
-                    archived: false
-                  });
+              scheduleMigration({
+                plan: scheduleMigrationPlan,
+                scheduleTime: this.state.dateTimeInput
+              }).then(() => {
+                fetchTransformationPlansAction({
+                  url: fetchTransformationPlansUrl,
+                  archived: false
                 });
-              } else if (this.state.startMigrationNow) {
-                scheduleMigrationNow(scheduleMigrationPlan.href);
-                toggleScheduleMigrationModal();
-              } else {
-                scheduleMigration({
-                  plan: scheduleMigrationPlan,
-                  scheduleTime: this.state.dateTimeInput
-                }).then(() => {
-                  fetchTransformationPlansAction({
-                    url: fetchTransformationPlansUrl,
-                    archived: false
-                  });
-                });
-              }
+              });
+              handleChange();
             }}
           >
             {__('Schedule')}
@@ -138,8 +70,6 @@ ScheduleMigrationModal.propTypes = {
   toggleScheduleMigrationModal: PropTypes.func,
   scheduleMigrationPlan: PropTypes.object,
   scheduleMigration: PropTypes.func,
-  scheduleMigrationNow: PropTypes.func,
-  scheduleCutover: PropTypes.func,
   fetchTransformationPlansAction: PropTypes.func,
   fetchTransformationPlansUrl: PropTypes.string
 };
